+++ babel.config.cjs
@@ -4,7 +4,5 @@ module.exports = {
         ['@babel/preset-react', {runtime: 'automatic'}],
         '@babel/preset-typescript',
     ],
-    plugins: [
-        'babel-plugin-dynamic-import-node'
-    ],
+    plugins: [],
 };