+++ src/components/form/socialButtonsComponent.tsx
@@ -11,6 +11,7 @@ import type { Provider, ProviderId } from '../../providers/providers';
 import { useReachfive } from '../../contexts/reachfive';
 
 import { Button, type ButtonProps } from './buttonComponent';
+import { useI18n } from '../../contexts/i18n';
 
 interface SocialButtonIconProps {
     className?: classes.Argument
@@ -52,6 +53,7 @@ interface SocialBtn extends ButtonProps {
 
 const SocialBtn = styled(Button).attrs((props: SocialBtn) => {
     const { provider } = props
+    const i18n = useI18n()
 
     return {
         themePrefix: 'socialButton',
@@ -59,7 +61,7 @@ const SocialBtn = styled(Button).attrs((props: SocialBtn) => {
         background: provider.btnBackgroundColor || provider.color,
         border: provider.btnBorderColor || provider.color,
         extendedClasses: classes(['r5-btn-social', `r5-btn-social-${provider.key}`]),
-        title: provider.name
+        title: i18n(`socialButton.${provider.key}.title`, undefined, () => provider.name)
     }
 })<SocialBtn>`
     margin-bottom: ${props => props.theme.spacing}px;
@@ -89,6 +91,7 @@ interface SocialButtonProps {
 
 const SocialButton = ({ provider, onClick, count }: SocialButtonProps) => {
     const theme = useTheme()
+    const i18n = useI18n()
 
     const inline = theme.socialButton.inline;
     const textVisible = theme.socialButton.textVisible;
@@ -106,7 +109,7 @@ const SocialButton = ({ provider, onClick, count }: SocialButtonProps) => {
         height={height}
         onClick={() => onClick(provider.key)}>
         <SocialButtonIcon icon={provider.icon} textVisible={textVisible} />
-        {textVisible && <SocialButtonText>{provider.name}</SocialButtonText>}
+        {textVisible && <SocialButtonText>{i18n(`socialButton.${provider.key}.title`, undefined, () => provider.name)}</SocialButtonText>}
     </SocialBtn>;
 };